procurement_auto_create_group: Set partner on created procurement group
