+++ procurement_auto_create_group/README.rst
@@ -84,6 +84,7 @@ Contributors
 * Jordi Ballester <jordi.ballester@forgeflow.com>
 * Lois Rilo <lois.rilo@forgeflow.com>
 * Héctor Villarreal Ortega <hector.villarreal@forgeflow.com>
+* Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 
 Maintainers
 ~~~~~~~~~~~
