+++ procurement_auto_create_group/__manifest__.py
@@ -1,13 +1,14 @@
 # Copyright 2017-2020 ForgeFlow, S.L.
+# Copyright 2021 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 {
     "name": "Procurement Auto Create Group",
-    "version": "14.0.1.0.0",
+    "version": "14.0.1.1.0",
     "development_status": "Production/Stable",
     "license": "AGPL-3",
     "summary": "Allows to configure the system to propose automatically new "
     "procurement groups during the procurement run.",
-    "author": "ForgeFlow," "Odoo Community Association (OCA)",
+    "author": "ForgeFlow, BCIM, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/stock-logistics-warehouse",
     "category": "Warehouse",
     "depends": ["stock"],
