+++ procurement_auto_create_group/models/procurement_group.py
@@ -1,8 +1,8 @@
 # Copyright 2017-2020 ForgeFlow, S.L.
+# Copyright 2021 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from odoo import _, api, models
-from odoo.exceptions import UserError
+from odoo import api, models
 
 
 class ProcurementGroup(models.Model):
@@ -10,25 +10,16 @@ class ProcurementGroup(models.Model):
 
     @api.model
     def _get_rule(self, product_id, location_id, values):
-        result = super()._get_rule(product_id, location_id, values)
+        rule = super()._get_rule(product_id, location_id, values)
         # If there isn't a date planned in the values it means that this
         # method has been called outside of a procurement process.
         if (
-            result
+            rule
             and not values.get("group_id")
-            and result.auto_create_group
+            and rule.auto_create_group
             and values.get("date_planned")
         ):
-            group_data = self._prepare_auto_procurement_group_data()
+            group_data = rule._prepare_auto_procurement_group_data()
             group = self.env["procurement.group"].create(group_data)
             values["group_id"] = group
-        return result
-
-    @api.model
-    def _prepare_auto_procurement_group_data(self):
-        name = self.env["ir.sequence"].next_by_code("procurement.group") or False
-        if not name:
-            raise UserError(_("No sequence defined for procurement group."))
-        return {
-            "name": name,
-        }
+        return rule
