+++ procurement_auto_create_group/readme/CONTRIBUTORS.rst
@@ -1,3 +1,4 @@
 * Jordi Ballester <jordi.ballester@forgeflow.com>
 * Lois Rilo <lois.rilo@forgeflow.com>
 * Héctor Villarreal Ortega <hector.villarreal@forgeflow.com>
+* Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
